--- server/routes/chat/summaryagent.ts
@@ -1,220 +1,0 @@
-import type { LlmConfig, DbRows, GeneDataTypeResult } from '#types'
-import { FILTER_TERM_DEFINITIONS, validate_filter } from './filter.ts'
-import {
-	formatTrainingExamples,
-	extractGenesetsFromPrompt,
-	validate_term,
-	buildCommonPrompt,
-	readJSONFile
-} from './utils.ts'
-import { route_to_appropriate_llm_provider } from './routeAPIcall.ts'
-import path from 'path'
-import fs from 'fs'
-
-type TermCategory = 'categorical' | 'float' | 'integer' | undefined
-
-// Default chart type for each (normalized cat1, normalized cat2) combination
-const CHILD_TYPE_DEFAULTS: Record<string, string> = {
-	'categorical:undefined': 'barchart',
-	'numeric:undefined': 'violin',
-	'categorical:categorical': 'barchart',
-	'numeric:categorical': 'violin',
-	'categorical:numeric': 'violin',
-	'numeric:numeric': 'sampleScatter'
-}
-
-// Overrides that are invalid (produce an error) for a given combination
-const CHILD_TYPE_INVALID: Record<string, Set<string>> = {
-	'categorical:undefined': new Set(['violin', 'boxplot', 'sampleScatter']),
-	'categorical:categorical': new Set(['violin', 'boxplot', 'sampleScatter'])
-}
-
-export async function extract_summary_terms(
-	prompt: string,
-	llm: LlmConfig,
-	dataset_db_output: { db_rows: DbRows[]; rag_docs: string[] },
-	dataset_json: any,
-	ds: any,
-	testing: boolean,
-	genesetNames: string[] = [],
-	geneFeatures: GeneDataTypeResult[],
-	aiFilesDir: string
-) {
-	const common_genes = geneFeatures.map(g => g.gene)
-	const Schema = {
-		$schema: 'http://json-schema.org/draft-07/schema#',
-		$ref: '#/definitions/SummaryType',
-		definitions: {
-			SummaryType: {
-				type: 'object',
-				properties: {
-					term: { type: 'string', description: 'Name of 1st term' },
-					term2: { type: 'string', description: 'Name of 2nd term' },
-					simpleFilter: {
-						type: 'array',
-						items: { $ref: '#/definitions/FilterTerm' },
-						description: 'Optional simple filter terms'
-					},
-					childType: {
-						type: 'string',
-						enum: ['violin', 'boxplot', 'sampleScatter', 'barchart'],
-						description:
-							'Optional explicit child type requested by the user. If omitted, the logic of the data types picks the child type.'
-					}
-				},
-				required: ['term', 'simpleFilter'],
-				additionalProperties: false
-			},
-			...FILTER_TERM_DEFINITIONS
-		}
-	} // This JSON schema is generated by ts-json-schema-generator. When SummaryType is updated, please update this schema by uncommenting the above code and running it locally
-	const matchedGenesets = extractGenesetsFromPrompt(prompt, genesetNames)
-
-	// Read Summary agent-specific JSON file
-	if (!fs.existsSync(path.join(aiFilesDir, 'summary.json')))
-		throw 'Summary agent file is not specified for dataset:' + ds.label
-	const summary_ds = await readJSONFile(path.join(aiFilesDir, 'summary.json'))
-	if (!summary_ds) throw 'Summary information is not present in the dataset file.'
-	if (summary_ds.TrainingData.length == 0) throw 'No training data is provided for the summary agent.'
-
-	const training_data = formatTrainingExamples(summary_ds.TrainingData)
-
-	const system_prompt =
-		'I am an assistant that extracts the summary terms from user query. The final output must be in the following JSON format with NO extra comments. The JSON schema is as follows: ' +
-		JSON.stringify(Schema) +
-		' term and term2 (if present) should ONLY contain single field names from the sqlite db — never comma-separated values, never category values like subtype names. ' +
-		'When the user says "across [field]" or "by [field]" (e.g. "across molecular subtypes", "by sex"), that field should be term2 — it is the grouping/overlay variable. ' +
-		'When the user mentions specific category VALUES by name (e.g. "between KMT2A and DUX4", "in male patients"), put those values in "simpleFilter" to restrict the data. ' +
-		'A query can have BOTH: e.g. "Vincristine sensitivity in male patients across molecular subtypes" → term=Vincristine sensitivity, term2=molecular subtype field, simpleFilter=[sex=male]. ' +
-		buildCommonPrompt({
-			ds,
-			dataset_json,
-			chart_ds: summary_ds,
-			dataset_db_output,
-			training_data,
-			common_genes,
-			prompt,
-			termFieldNames: ['term', 'term2'],
-			matchedGenesets
-		})
-
-	const response: string = await route_to_appropriate_llm_provider(system_prompt, llm)
-	if (testing) {
-		// When testing, send raw LLM response
-		const test_response = JSON.parse(response)
-		test_response.plot = 'summary'
-		test_response.type = 'plot'
-		return test_response
-	} else {
-		// In actual production (inside PP) send LLM output for validation
-		return validate_summary_response(response, ds, geneFeatures)
-	}
-}
-
-function validate_summary_response(response: string, ds: any, geneFeatures: GeneDataTypeResult[]) {
-	const response_type = JSON.parse(response)
-	const pp_plot_json: any = { chartType: 'summary' }
-	let text = ''
-	if (response_type.text) text = response_type.text
-	if (!response_type.term) {
-		text += 'term type is not present in summary output'
-		return { type: 'text', text: text }
-	}
-	const term1_validation = validate_term(response_type.term, ds, geneFeatures)
-	if (term1_validation.text.length > 0) {
-		text += term1_validation.text
-		return { type: 'text', text: text }
-	} else {
-		pp_plot_json.term = term1_validation.term_type
-		if (term1_validation.category == 'float' || term1_validation.category == 'integer') {
-			pp_plot_json.term.q = { mode: 'continuous' }
-		}
-		pp_plot_json.category = term1_validation.category
-	}
-
-	if (response_type.term2) {
-		const term2_validation = validate_term(response_type.term2, ds, geneFeatures)
-		if (term2_validation.text.length > 0) {
-			text += term2_validation.text
-			return { type: 'text', text: text }
-		} else {
-			pp_plot_json.term2 = term2_validation.term_type
-			if (term2_validation.category == 'float' || term2_validation.category == 'integer') {
-				pp_plot_json.term2.q = { mode: 'continuous' }
-			}
-			pp_plot_json.category2 = term2_validation.category
-		}
-	}
-	/** Based on data types of term and term2, decide the most appropriate chart type.
-	 *  The user can override the default by explicitly mentioning a chart type in their prompt,
-	 *  which the LLM parses into the "childType" field. Invalid overrides produce an error. */
-	const llmChildType =
-		response_type.childType && ['violin', 'boxplot', 'sampleScatter', 'barchart'].includes(response_type.childType)
-			? response_type.childType
-			: undefined
-	const resolved = resolveChildType(pp_plot_json.category, pp_plot_json.category2, llmChildType)
-
-	if (resolved.error) {
-		text += resolved.error
-		return { type: 'text', text: text }
-	} else {
-		pp_plot_json.childType = resolved.childType
-		// For two numeric variables displayed as violin/boxplot, discretize term2
-		if (pp_plot_json.childType == 'barchart') {
-			pp_plot_json.term.q = { mode: 'discrete' }
-			if (pp_plot_json.term2) {
-				pp_plot_json.term2.q = { mode: 'discrete' }
-			}
-		}
-		if (resolved.bothNumeric && (resolved.childType == 'violin' || resolved.childType == 'boxplot')) {
-			pp_plot_json.term2.q = { mode: 'discrete' }
-		}
-	}
-
-	delete pp_plot_json.category
-	if (pp_plot_json.category2) delete pp_plot_json.category2
-
-	if (response_type.simpleFilter && response_type.simpleFilter.length > 0) {
-		const validated_filters = validate_filter(response_type.simpleFilter, ds, '')
-		if (validated_filters.text.length > 0) {
-			text += validated_filters.text
-			return { type: 'text', text: text }
-		} else {
-			pp_plot_json.filter = validated_filters.simplefilter
-		}
-	}
-	return { type: 'plot', plot: pp_plot_json }
-}
-
-function resolveChildType(
-	cat1: TermCategory,
-	cat2: TermCategory,
-	llmChildType: string | undefined
-): { childType?: string; error?: string; bothNumeric?: boolean } {
-	const norm1 = cat1 == 'float' || cat1 == 'integer' ? 'numeric' : cat1 || 'undefined'
-	const norm2 = cat2 == 'float' || cat2 == 'integer' ? 'numeric' : cat2 || 'undefined'
-	const key = norm1 + ':' + norm2
-	const defaultType = CHILD_TYPE_DEFAULTS[key]
-	if (!defaultType) {
-		// Unknown category combination — should not happen, fall back to barchart
-		return { childType: 'barchart' }
-	}
-	const invalid = CHILD_TYPE_INVALID[key]
-
-	if (llmChildType && invalid && invalid.has(llmChildType)) {
-		return {
-			error:
-				'Invalid plot type supplied by the user: ' +
-				llmChildType +
-				'. For ' +
-				key.replace(':', ' and ') +
-				' variables the plot type should always be ' +
-				defaultType
-		}
-	}
-
-	return {
-		childType: llmChildType || defaultType,
-		bothNumeric: norm1 == 'numeric' && norm2 == 'numeric'
-	}
-}
