+++ server/src/app.routes.js
@@ -46,8 +46,6 @@ export const routeFiles = [
 	import('../routes/termdb.diffMeth.ts'),
 	import('../routes/termdb.violinBox.ts'),
 	import('../routes/termdb.chat.ts'),
-	//import('../routes/termdb.chat2.ts'),
-	import('../routes/termdb.chat3.ts'),
 	import('../routes/termdb.categories.ts'),
 	import('../routes/termdb.cluster.ts'),
 	import('../routes/termdb.cohort.summary.ts'),
